+++ changelog.md
@@ -5,6 +5,10 @@
 
 # ChangeLog:
 
+#### v11.01
+
+- Fixed bugs
+
 #### v11.00
 
 - Updated the command to execute snape scripts